refactor(lsp): move stdio server from wado-cli into wado-lsp (WEP-2026-04-18)
Move the LSP server implementation (Content-Length framing, JSON-RPC
dispatch, wire types, FilesystemCompilerHost core) from wado-cli into
wado-lsp so the same Rust code builds for both native and wasm32-wasip2
per the LSP architecture WEP.

- wado-lsp gains `src/server/{mod,transport,dispatch,rpc}.rs`,
  `src/host.rs`, and a `wado-lsp` binary.
- Transport is synchronous via `std::io` (tokio's `io-std` is unavailable
  on wasm32-*); dispatch stays async to await Engine queries.
- wado-cli `compiler_host` now wraps the wado-lsp host with CLI-only
  decorations (timestamps, log-level filter, stderr printing).
- wado-cli `lsp` subcommand is a thin delegator; `lsp_adapter.rs` and
  `lsp_rpc.rs` are deleted.
- rust-toolchain adds `wasm32-wasip2`; mise gains
  `build-wado-lsp-wasm` / `watch-wado-lsp-wasm`.
