+++ rust-toolchain.toml
@@ -1,4 +1,4 @@
 [toolchain]
 channel = "1.95"
 components = ["clippy", "rustfmt"]
-targets = ["wasm32-unknown-unknown"]
+targets = ["wasm32-unknown-unknown", "wasm32-wasip2"]
