+++ wado-lsp/src/host.rs
@@ -1,0 +1,66 @@
+//! Filesystem-based `CompilerHost` for embedders that need silent diagnostic
+//! collection plus relative-path source loading.
+//!
+//! This is the default host used by the LSP server. Consumers that additionally
+//! want to decorate output (timestamps, log-level filtering, stderr printing)
+//! wrap this host.
+
+use std::path::PathBuf;
+use std::sync::Mutex;
+
+use wado_compiler::{CompilerHost, Diagnostic, Severity, SourceError};
+
+#[derive(Debug)]
+pub struct FilesystemCompilerHost {
+    base_path: PathBuf,
+    diagnostics: Mutex<Vec<Diagnostic>>,
+}
+
+impl FilesystemCompilerHost {
+    #[must_use]
+    pub fn new(base_path: PathBuf) -> Self {
+        Self {
+            base_path,
+            diagnostics: Mutex::new(Vec::new()),
+        }
+    }
+
+    pub fn base_path(&self) -> &PathBuf {
+        &self.base_path
+    }
+
+    pub fn diagnostics(&self) -> Vec<Diagnostic> {
+        self.diagnostics.lock().unwrap().clone()
+    }
+
+    pub fn has_errors(&self) -> bool {
+        self.diagnostics
+            .lock()
+            .unwrap()
+            .iter()
+            .any(|d| d.severity == Severity::Error)
+    }
+
+    /// Append a diagnostic to the collected buffer without emitting it.
+    ///
+    /// Wrappers call this after performing their own side effects (e.g. stderr
+    /// printing) so the buffer remains the single source of truth for
+    /// `has_errors` / `diagnostics()`.
+    pub fn collect_diagnostic(&self, diagnostic: Diagnostic) {
+        self.diagnostics.lock().unwrap().push(diagnostic);
+    }
+}
+
+impl CompilerHost for FilesystemCompilerHost {
+    async fn load_source(&self, path: &str) -> Result<Vec<u8>, SourceError> {
+        let full_path = self.base_path.join(path);
+        std::fs::read(&full_path).map_err(|e| SourceError::IoError {
+            path: full_path.display().to_string(),
+            message: e.to_string(),
+        })
+    }
+
+    fn emit_diagnostic(&self, diagnostic: Diagnostic) {
+        self.collect_diagnostic(diagnostic);
+    }
+}
